submitted patch for issue # 19 (speaker profile / listing) with major modifications.
(I couldn't test the TempData in the controllers, as it gets a little more in depth.  I'll work on that for the next commit, but I wanted everyone to be able to get the latest bits).

Thanks for the patch!

--HG--
extra : convert_revision : svn%3A25cf9c4d-fa41-0410-a78d-8fe958b11fb0/trunk%4087